refactor: :art: make code blocks their own class
They have their own class and stuff
